Move `ReceiveFuture` into separate top-level module
