+++ src/context.rs
@@ -6,13 +6,13 @@ use std::task::Poll;
 use std::{mem, task};
 
 use futures_core::future::BoxFuture;
-use futures_core::FusedFuture;
 use futures_util::future::{self, Either};
 use futures_util::FutureExt;
 
 use crate::envelope::{Shutdown, Span};
-use crate::inbox::rx::{ReceiveFuture as InboxReceiveFuture, RxStrong};
+use crate::inbox::rx::RxStrong;
 use crate::inbox::ActorMessage;
+use crate::recv_future::{Message, ReceiveFuture};
 use crate::{inbox, Actor, Address, Error, WeakAddress};
 
 /// `Context` is used to control how the actor is managed and to get the actor's address from inside
@@ -310,40 +310,6 @@ impl<A> Clone for Context<A> {
             running: self.running,
             mailbox: self.mailbox.clone(),
         }
-    }
-}
-
-/// A message sent to a given actor, or a notification that it should shut down.
-pub struct Message<A>(pub(crate) ActorMessage<A>);
-
-/// A future which will resolve to the next message to be handled by the actor.
-///
-/// # Cancellation safety
-///
-/// This future is cancellation-safe in that no messages will ever be lost, even if this future is
-/// dropped half-way through. However, reinserting the message into the mailbox may mess with the
-/// ordering of messages and they may be handled by the actor out of order.
-///
-/// If the order in which your actors process messages is not important to you, you can consider this
-/// future to be fully cancellation-safe.
-///
-/// If you wish to maintain message ordering, you can use [`FutureExt::now_or_never`] to do a final
-/// poll on the future. [`ReceiveFuture`] is guaranteed to complete in a single poll if it has
-/// remaining work to do.
-#[must_use = "Futures do nothing unless polled"]
-pub struct ReceiveFuture<A>(InboxReceiveFuture<A, RxStrong>);
-
-impl<A> Future for ReceiveFuture<A> {
-    type Output = Message<A>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Self::Output> {
-        self.0.poll_unpin(cx).map(Message)
-    }
-}
-
-impl<A> FusedFuture for ReceiveFuture<A> {
-    fn is_terminated(&self) -> bool {
-        self.0.is_terminated()
     }
 }
